Up the amount of allowed messages by 1 in all timeframes
This is mostly just for the short counter, but then it also makes sense
to increase the other limits. Basically there's a bug in MC where for
some reason the client will repeat the same message twice. This will
without fail trigger the short counter limit.

With this change players are now allowed to send up to and including 2
messages within the short time limit (.4 seconds,) which should prevent
people from getting muted when this bug occurs.

It might be worthwhile in the future to change it so that if a
message is repeated within .4 seconds, the second copy will be canceled.